Various fixs to Travis CI build.
 - Fix a random build failure where Riak would still be setting up
Riak Search when the search test is run, resulting in a failure.
Instead wait for Riak to have started the search service.  Also
wait for the kv service to start, in case that ever becomes a
problem in the future.
 - Instead of manually specifying the test dependencies, use the -t
option to go get.
 - Build/test all packages in the repository.  While they already
should have been, this ensures it in the future.